backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# Celery
celery==5.3.6
celery[redis]==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Date/Time
python-dateutil==2.8.2

# Analytics and Data Processing
pandas==2.2.0
numpy==1.26.3

# Forecasting
prophet==1.1.5
statsmodels==0.14.1

# Export
openpyxl==3.1.2
reportlab==4.0.9
xlsxwriter==3.1.9

# Telegram Bot
python-telegram-bot==20.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0
